requirements_2023.txt
amqp==1.4.8
anyjson==0.3.3
asgiref==3.6.0
asn1crypto==1.5.1
Babel==1.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.22
bleach==1.4
boto==2.38.0
celery==3.1.19
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==2.4.2
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==0.3.0
dj-stripe==1.1.0
Django==2.2.28
django-ajax==0.3.0
django-allauth==0.40.0
django-appconf==1.0.5
django-braces==1.12.0
django-cacheops==5.0.1
django-celery==3.3.1
django-compat==1.0.15
django-compressor==1.4
django-countries==7.5.1
django-crispy-forms==1.11.0
django-extensions==1.5.2
django-extra-views==0.11.0
django-filer==2.2.4
django-filter==2.3.0
django-floppyforms==1.9.0
django-haystack==2.8.1
django-hijack==2.1.10
django-image-cropping==1.7
django-js-asset==2.0.0
django-localflavor==1.1
django-model-utils==2.0
django-mptt==0.14.0
django-multi-email-field==0.3.1
django-oscar==2.0
django-oscar-paypal==2.0.0
django-oscar-stripe-sca==0.1
django-phonenumber-field==2.0.1
django-pipeline==1.5.0
django-polymorphic==3.1.0
django-redis-cache==1.6.3
django-sortedm2m==3.0.2
django-storages-redux==1.3
django-tables2==1.21.2
django-thumbor==0.5.3
django-tinymce==3.5.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.1
django-wkhtmltopdf==3.2.0
djangorestframework==3.13.0
djrill==2.1.0
dnspython==1.15.0
dnspython3==1.15.0
docopt==0.4.0
easy-thumbnails==2.8.5
elasticsearch==1.7.0
email-validator==1.0.1
factory-boy==2.12.0
Faker==14.2.1
funcy==1.18
futures==3.0.5
gunicorn==19.3.0
html5lib==0.999
humanfriendly==1.32
idna==3.4
importlib-resources==5.4.0
ipython==3.1.0
jsonfield==2.0.2
kombu==3.0.30
libsass==0.7.0
libthumbor==1.0.1
lxml==4.9.2
mailchimp==2.0.9
metrics==0.3.3
mock==1.0.1
MonthDelta==1.0b0
natsort==3.5.0
newrelic==2.54.0.41
oauthlib==0.7.2
pathlib2==2.3.7.post1
pathspec==0.5.5
paypalrestsdk==1.13.1
phonenumbers==8.13.7
Pillow==9.4.0
psycopg2-binary==2.8.6
purl==1.0.3
pycountry==1.10
pycparser==2.21
pycrypto==2.6.1
pycurl==7.45.2
pyelasticsearch==1.4
Pygments==2.2.0
pyOpenSSL==16.2.0
pystache==0.6.0
python-dateutil==2.4.2
python3-openid==3.2.0
pytz==2015.2
raven==5.7.2
redis==2.10.6
reportlab==3.6.12
requests==2.27.1
requests-oauthlib==0.4.2
signals==0.0.2
simplejson==3.18.4
six==1.16.0
smallslive-metrics-app @ git+https://github.com/SmallsLIVE/smallslive-metrics-djangoapp.git@e5b4b537a4356ea4e2ba1e9a20409d33c8fef544
sorl-thumbnail==11.12.1b0
sqlparse==0.4.3
static==1.1.1
static3==0.5.1
stripe==5.2.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.1.1
Unidecode==0.4.17
urllib3==1.26.15
webencodings==0.5.1
whitenoise==2.0.4
wkhtmltopdf-pack-ng==0.12.3.0
XlsxWriter==0.7.3
zipp==3.6.0
